Bug 75: linker script workaround for newlib heap corruption
There appears to be cryptic bug in newlib heap allocator which causes it to
fail when the first malloc request fits exactly (or within 4 bytes) of
the next 4k page boundary form the heap base.

For example, heap base(HOFFSET) = 0xe40, 1st malloc size(SZ) = 0x1ac
then, 4096 - (HOFFSET + SZ) = 4.

If the heap base is not aligned to 16-byte boundary, it forces an extra
alignment step and _sbrk() call, which averts the issue. The root cause
is not fully understood but this patch works around the problem by
changed causing the initial heap base to be mis-aligned.

Change-Id: I71efa04b744be46acd6e4e1e8d1ac361f3a063df