[nanoMIPS] Adjust uhi linker scripts for lld
There are differences in the implementation of lld and gold linker
scripts and their parsing. This patch addresses one of them, that would
be hard to fix in the linker itself.

Description:

In the uhi linker script, this can be found:
EXTERN(__exception_handle_verbose)

...

PROVIDE(__exception_handle = (DEFINED(__exception_handle_quiet)
                                     ? __exception_handle_quiet
                                     : __exception_handle_verbose));

gold takes __exception_handle_verbose as the value of
__exception_handle while lld takes __exception_handle_quiet

lld tries to resolve every symbol that is referenced in the linker
script, so it sees that __exception_handle_quiet is referenced
(after the ? operator) and tries to find a symbol to match it.
lld finds this symbol in libhal.a:mips_excpt_handler_quiet.o and
resolves it. This means whenever __exception_handle_quiet is
referenced and libhal.a is used in linking, quiet exception handling
will be the exception handling behaviour.

gold also finds the symbol, but because it is not defined as EXTERN,
nor referenced by some other object file, it is treated as undefined
and it does not end up in the output file.

This is solved by not mentioning the quiet exception handling at all.
So to specify exception handling users need to change the verbose
handle to quiet handle on two places: the one in EXTERN and one in
PROVIDE.

Change-Id: I70cfa08f37ef32f70452e990e929e4a9ff6ab7a7